Remove duplicated package.
