+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/internal/offline/repository/database/converter/internal/VoteConverter.kt
@@ -14,28 +14,46 @@
  * limitations under the License.
  */
 
-package io.getstream.chat.android.client.internal.offline.repository.database.database.converter.internal
+package io.getstream.chat.android.client.internal.offline.repository.database.converter.internal
 
 import androidx.room.TypeConverter
 import com.squareup.moshi.adapter
-import io.getstream.chat.android.client.internal.offline.repository.domain.user.internal.PrivacySettingsEntity
+import io.getstream.chat.android.client.internal.offline.repository.domain.message.internal.VoteEntity
 
-internal class PrivacySettingsConverter {
+internal class VoteConverter {
 
     @OptIn(ExperimentalStdlibApi::class)
-    private val entityAdapter = moshi.adapter<PrivacySettingsEntity>()
+    private val entityAdapter = moshi.adapter<VoteEntity>()
+
+    @OptIn(ExperimentalStdlibApi::class)
+    private val entityListAdapter = moshi.adapter<List<VoteEntity>>()
 
     @TypeConverter
-    fun stringToPrivacySettings(data: String?): PrivacySettingsEntity? {
+    fun stringToVote(data: String?): VoteEntity? {
         return data?.let {
             entityAdapter.fromJson(it)
         }
     }
 
     @TypeConverter
-    fun privacySettingsToString(entity: PrivacySettingsEntity?): String? {
+    fun voteToString(entity: VoteEntity?): String? {
         return entity?.let {
             entityAdapter.toJson(it)
+        }
+    }
+
+    @TypeConverter
+    fun stringToVoteList(data: String?): List<VoteEntity>? {
+        if (data.isNullOrEmpty() || data == "null") {
+            return emptyList()
+        }
+        return entityListAdapter.fromJson(data)
+    }
+
+    @TypeConverter
+    fun voteListToString(entities: List<VoteEntity>?): String? {
+        return entities?.let {
+            entityListAdapter.toJson(it)
         }
     }
 }
