+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/internal/offline/repository/database/converter/ExtraDataConverterTest.kt
@@ -16,7 +16,7 @@
 
 package io.getstream.chat.android.client.internal.offline.repository.database.converter
 
-import io.getstream.chat.android.client.internal.offline.repository.database.database.converter.internal.ExtraDataConverter
+import io.getstream.chat.android.client.internal.offline.repository.database.converter.internal.ExtraDataConverter
 import io.getstream.chat.android.client.test.utils.TestDataHelper
 import io.getstream.chat.android.test.TestLoggingHelper
 import io.getstream.chat.android.test.multiMapOf
