+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/internal/offline/repository/database/converter/MapConverterTest.kt
@@ -16,7 +16,7 @@
 
 package io.getstream.chat.android.client.internal.offline.repository.database.converter
 
-import io.getstream.chat.android.client.internal.offline.repository.database.database.converter.internal.MapConverter
+import io.getstream.chat.android.client.internal.offline.repository.database.converter.internal.MapConverter
 import io.getstream.chat.android.client.internal.offline.repository.domain.channel.member.internal.MemberEntity
 import io.getstream.chat.android.client.internal.offline.repository.domain.channel.userread.internal.ChannelUserReadEntity
 import io.getstream.chat.android.client.test.utils.TestDataHelper
