+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/internal/offline/repository/database/converter/SetConverterTest.kt
@@ -16,7 +16,7 @@
 
 package io.getstream.chat.android.client.internal.offline.repository.database.converter
 
-import io.getstream.chat.android.client.internal.offline.repository.database.database.converter.internal.SetConverter
+import io.getstream.chat.android.client.internal.offline.repository.database.converter.internal.SetConverter
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.Test
